Default fullscreen + fixing some warning errors
Now the editor is open in fullscreen by default

+++ Ilargi/source/Base/Input.cpp
@@ -28,6 +28,6 @@ namespace Ilargi
 	{
 		double x, y;
 		glfwGetCursorPos(window, &x, &y);
-		return vec2(x, y);
+		return vec2((float)x, (float)y);
 	}
 }
